FINALLY fixed Story loot
The answer was ObjectCategory. A lot of the erroneous attempts were undone. Arena items were also overwritten to force them out of the treasure table, as well. Arena play is 98% not affected.
Fixed Ult. Incarnate summon having a Memory cost.
Added a sorta cheat-y Restoration skill so a monster can heal with its own magic ability and not potions and weaker skills.
Added placeholder string for the embiggened Slane in Arena mode.